+++ src/fx/consume_tick.rs
@@ -45,7 +45,7 @@ impl Shader for ConsumeTick {
 #[cfg(test)]
 #[cfg(feature = "dsl")]
 mod tests {
-    use crate::{fx, Shader};
+    use crate::fx;
 
     #[test]
     fn consume_tick() {