+++ src/render_effect.rs
@@ -1,23 +1,20 @@
 use ratatui::{buffer::Buffer, layout::Rect, Frame};
 
-use crate::{shader::Shader, Duration};
+use crate::{Duration, Effect};
 
 pub trait EffectRenderer<T> {
     fn render_effect(&mut self, effect: &mut T, area: Rect, last_tick: Duration);
 }
 
-impl<S: Shader> EffectRenderer<S> for Frame<'_> {
-    fn render_effect(&mut self, effect: &mut S, area: Rect, last_tick: Duration) {
-        render_effect(effect, self.buffer_mut(), area, last_tick);
+impl EffectRenderer<Effect> for Frame<'_> {
+    fn render_effect(&mut self, effect: &mut Effect, area: Rect, last_tick: Duration) {
+        // render_effect(effect, self.buffer_mut(), area, last_tick);
+        effect.process(last_tick, self.buffer_mut(), area);
     }
 }
 
-impl<S: Shader> EffectRenderer<S> for Buffer {
-    fn render_effect(&mut self, effect: &mut S, area: Rect, last_tick: Duration) {
-        render_effect(effect, self, area, last_tick);
+impl EffectRenderer<Effect> for Buffer {
+    fn render_effect(&mut self, effect: &mut Effect, area: Rect, last_tick: Duration) {
+        effect.process(last_tick, self, area);
     }
-}
-
-fn render_effect<S: Shader>(effect: &mut S, buf: &mut Buffer, area: Rect, last_tick: Duration) {
-    effect.process(last_tick, buf, area);
 }